Minor bug fixes and one debug statement.
The "Download" interface seems to use a value of "zero" to mean "unlimited mode" and "-1" to mean set download rate to zero. This causes a few issues when trying to set the active torrent upload/download limits to zero before starting the test. Specifically the "DownloadImpl.setMaximumDownloadKBPerSecond" method does not handle this condition well. So instead a new "setDownloadRateLimitBytesPerSecond" method is used and a debug statement is added to the KBPS method. The debug statement is to warn any plug-in writers that might be trying to use that method.