Fix #1: Store referenced downloads in keyed data for the torrent menu item, since the setData(Object) slot seems to be used by something else (was getting ClassCastExceptions).
Fix #2: Wrapped the call to set the menu item data to be in a SWT thread.

Fix #3: We record whether or not the current view is the "Show Details" view, and store that on the menu item - so we can then decide whether it is appropriate to have a "Show Details" menu option (rather than having it always disabled on the Torrent menu).