Phase #1 of my changes to make download property listeners to be more lightweight: the plugin API now defines the DownloadAttributeListener interface. This replaces the now deprecated DownloadPropertyListener interface.
When plugins register listeners, they will have to indicate which torrent attribute they are interested in, and whether they are interested in the attribute being changed (99% of the time), or whether they are intercepting the attribute as it is about to be read (that Team Seeder plugin).

Phase #1 is providing the API at the plugin level - internally, we just register property listeners and provide a bridge between the two. Phase #2 is where we restructure the listeners at the DownloadManagerState level (and remove the "property" listener equivalent in the core). Phase #3 is where all property listener support is dropped (apart from doing some specialised stuff for backwards compatibility).