*Phew*... Jython's core doesn't support multiple concurrent interpreters as well as I wished it could.
In particular, there's a static reference to sys.stdout and sys.stderr - it's shared across interpreters. This means as soon as you opened up a new interpreter, all output (even from other interpreters) was being displayed in the recently opened interpreter.

To workaround this, we have to handle this in a special way - stderr and stdout will map to a PrintWriter, which turns the string into an object which identifies what type of output it is (stdout or stderr) and what thread added it. That then gets passed to a "multisource data sink", which passes it to every console for consideration.

As each console has its own execution thread, it compares the thread recorded with its own thread - displaying it if the threads are the same, ignoring it if the thread belongs to another console.

If the text has come from some other unknown thread (like if you were printing from an Azureus event dispatch thread), consoles will print it, but in italic, to indicate that the output might not be intended for this particular console.