PluginConfigSource - I spent ages writing this code (and all the other code to support it).
How it works - we provide an API to allow external objects that there is a parameter which we may want to monitor (if it appears as a plugin parameter with the same config key prefix as we were expecting). The PluginConfigImpl is the only class that does reporting of that at the moment.

Once we encounter a new parameter to monitor, we add a parameter listener to copy across (or maybe remove) the setting from a map that we keep. The ConfigurationManager instance will keep the value setting too - but we register it as a transient parameter so it doesn't get saved.

We then use the resilient file stuff to save the values with bencoding / bdecoding.