Modified these files so that they get values via PluginConfig.getUnsafeParameter (easier for me to hook into when parameters get used).
Some of the implementations of methods here were a bit overcomplicated (some of the methods for BooleanParameterImpl, for example), so I made that a bit more straight forward.

All the parameters need to call a method on PluginConfigImpl to register their existence (so that the external config source objects can pick up on it).