Added a new API for plugins to listen to log alerts - I didn't like the old API, because it was limited (you couldn't pass additional information without adding a new listener interface).
So, for example, timeout values weren't passed to listeners. And even if we added a new type of listener to support that, what about if we extend alerts to include custom icon types? Or define "categories" of alerts?

Therefore we present a plugin equivalent of LogEvent, so we can extend the object without breaking or adding additional method signatures.