OK, this was on my to-do list for a while, but if we're going to fix it, let's change it properly. This is one of the last instances of us using a property listener rather than an attribute listener internally - the reason it didn't get converted with the other code is because it didn't specifically refer to any attribute that it was interested in (i.e. it was potentially interested in *all* of them).
We definitely want to propogate the category attribute over from the download. The "share properties" attribute never exists as a normal download attribute, so we don't have to listen for that. So we'll just change this to listen to the category attribute only. If this breaks for some reason, revert back to the code in 1.13, and turn off the plugin debug warning.

Note - if the plugin debug warning stuff is getting in the way, you can just modify the "behaviour" constants in the main class (in the static block).